+++ src/main/java/net/brnbrd/delightful/common/item/IConfigured.java
@@ -3,7 +3,6 @@ package net.brnbrd.delightful.common.item;
 import net.brnbrd.delightful.Util;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.ItemLike;
 import java.util.List;
 
@@ -19,9 +18,5 @@ public interface IConfigured extends ItemLike {
             return false;
         }
         return true;
-    }
-
-    default ItemStack getCreative() {
-        return this.enabled() ? this.asItem().getDefaultInstance() : ItemStack.EMPTY;
     }
 }