+++ src/main/java/net/brnbrd/delightful/common/item/knife/compat/twilightforest/IronwoodKnifeItem.java
@@ -1,5 +1,6 @@
 package net.brnbrd.delightful.common.item.knife.compat.twilightforest;
 
+import net.brnbrd.delightful.Util;
 import net.brnbrd.delightful.common.item.DelightfulItems;
 import net.brnbrd.delightful.common.item.DelightfulTiers;
 import net.brnbrd.delightful.common.item.knife.CompatKnifeItem;
@@ -7,6 +8,7 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.RecipeType;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import vectorwing.farmersdelight.common.registry.ModEnchantments;
 
@@ -21,9 +23,7 @@ public class IronwoodKnifeItem extends CompatKnifeItem {
   }
 
   @Override
-  public ItemStack getCreative() {
-    ItemStack knife = super.getCreative();
-    knife.enchant(ModEnchantments.BACKSTABBING.get(), 1);
-    return knife;
+  public @NotNull ItemStack getDefaultInstance() {
+    return Util.enchant(super.getDefaultInstance(), ModEnchantments.BACKSTABBING.get(), 1);
   }
 }